+++ RingPublishingGDPR/src/main/java/com/ringpublishing/gdpr/internal/cmp/CmpWebViewAction.java
@@ -1,11 +1,12 @@
 package com.ringpublishing.gdpr.internal.cmp;
 
 import android.text.TextUtils;
-import android.util.Log;
 import android.webkit.CookieManager;
 
+import com.ringpublishing.gdpr.RingPublishingGDPRError;
 import com.ringpublishing.gdpr.RingPublishingGDPRListener;
 import com.ringpublishing.gdpr.internal.cmp.CmpAction.ActionType;
+import com.ringpublishing.gdpr.internal.log.Logger;
 import com.ringpublishing.gdpr.internal.storage.Storage;
 import com.ringpublishing.gdpr.internal.view.FormViewImpl;
 
@@ -15,9 +16,6 @@ import androidx.annotation.NonNull;
 
 public class CmpWebViewAction implements CmpWebViewActionCallback
 {
-
-    private final String TAG = CmpWebViewAction.class.getCanonicalName();
-
     @NonNull
     private final Storage storage;
 
@@ -37,27 +35,29 @@ public class CmpWebViewAction implements CmpWebViewActionCallback
     @Override
     public void onActionLoaded()
     {
-        Log.i(TAG, "Cmp site is ready");
+        Logger.get().info( "Cmp site is ready onActionLoaded");
         formViewImpl.post(formViewImpl::cmpReady);
     }
 
     @Override
     public void onActionComplete()
     {
+        Logger.get().info( "Cmp site onActionComplete");
         formViewImpl.formSubmittedAction();
     }
 
     @Override
     public void onActionError(String error)
     {
-        Log.w(TAG, "Error: " + error);
         if (formViewImpl.isOnline())
         {
-            closeForm();
+            closeFormByError("Error when try execute action in webview. Error: " + error);
         }
         else
         {
+            Logger.get().error("User is offline. Error when try execute action in webview. Error: " + error);
             formViewImpl.showError();
+            ringPublishingGDPRListener.onError(RingPublishingGDPRError.WEBVIEW_LOADING_FAIL, "User is offline. Error when try execute action in webview. Error: " + error);
         }
     }
 
@@ -68,18 +68,21 @@ public class CmpWebViewAction implements CmpWebViewActionCallback
         {
             try
             {
+                Logger.get().info("Try to save TCData from javascript");
                 storage.saveTCData(tcData);
+
             }
             catch (JSONException e)
             {
                 storage.clearAllConsentData();
-                Log.e(TAG, "saveTCData fail!!", e);
+                closeFormByError("When receive tcData from javascript and want to save TCData, " +
+                        "then have JSONException error:" + e.getLocalizedMessage() + " TcData to parse is: " + tcData);
             }
         }
         else
         {
             storage.clearAllConsentData();
-            Log.e(TAG, "Save TCData fail");
+            closeFormByError("When receive tcData from javascript, result is not success. TcData is: " + tcData);
         }
 
         CookieManager.getInstance().flush();
@@ -100,31 +103,46 @@ public class CmpWebViewAction implements CmpWebViewActionCallback
             try
             {
                 storage.saveConsentData(dlData);
+
+                boolean closeForm = formViewImpl.waitingActionFinish(ActionType.GET_COMPLETE_CONSENT_DATA);
+                if (closeForm)
+                {
+                    closeForm();
+                    ringPublishingGDPRListener.onConsentsUpdated();
+                }
             }
             catch (JSONException e)
             {
                 storage.clearAllConsentData();
-                Log.e(TAG, "Fail saving consent data", e);
+                Logger.get().error("Fail saving consent data" + e.getLocalizedMessage());
+                closeFormByError("When try to save consent data, then have JSONException with error: "
+                        + e.getLocalizedMessage() + " Parsed data is: " + dlData);
             }
         }
         else
         {
             storage.clearAllConsentData();
-            Log.e(TAG, "Save dlData fail");
+            Logger.get().error("Save dlData fail.Error " + error);
+            closeFormByError("When try to get complete consent data, then have error: " + error + " dlData:" + dlData);
         }
 
         CookieManager.getInstance().flush();
+    }
 
-        boolean closeForm = formViewImpl.waitingActionFinish(ActionType.GET_COMPLETE_CONSENT_DATA);
-        if (closeForm)
+    private void closeFormByError(String errorMessage)
+    {
+        Logger.get().error("Close form by error. Error message: " + errorMessage);
+
+        if (ringPublishingGDPRListener != null)
         {
-            closeForm();
-            ringPublishingGDPRListener.onConsentsUpdated();
+            ringPublishingGDPRListener.onError(RingPublishingGDPRError.CLOSE_FORM_WITH_ERROR, errorMessage);
         }
+        closeForm();
     }
 
     private void closeForm()
     {
+        Logger.get().info("Cmp closeForm");
         storage.saveLastAPIConsentsCheckStatus(null);
         storage.setConsentOutdated(false);
         formViewImpl.hideForm();