+++ RingPublishingGDPR/src/main/java/com/ringpublishing/gdpr/internal/task/FetchConfigurationTask.java
@@ -1,11 +1,10 @@
 package com.ringpublishing.gdpr.internal.task;
 
-import android.util.Log;
-
 import com.ringpublishing.gdpr.RingPublishingGDPRError;
 import com.ringpublishing.gdpr.RingPublishingGDPRListener;
 import com.ringpublishing.gdpr.internal.api.Api;
 import com.ringpublishing.gdpr.internal.api.Api.ConfigurationCallback;
+import com.ringpublishing.gdpr.internal.log.Logger;
 import com.ringpublishing.gdpr.internal.model.RequestsState;
 import com.ringpublishing.gdpr.internal.model.TenantConfiguration;
 import com.ringpublishing.gdpr.internal.model.TenantState;
@@ -16,8 +15,6 @@ import androidx.annotation.Nullable;
 
 public class FetchConfigurationTask
 {
-
-    private final String TAG = FetchConfigurationTask.class.getCanonicalName();
 
     @NonNull
     private final Api api;
@@ -33,6 +30,8 @@ public class FetchConfigurationTask
 
     @NonNull
     private final RingPublishingGDPRListener ringPublishingGDPRListener;
+
+    private final Logger log = Logger.get();
 
     public FetchConfigurationTask(@NonNull Api api,
                                   @NonNull Storage storage,
@@ -54,26 +53,27 @@ public class FetchConfigurationTask
             @Override
             public void onSuccess(@NonNull String url, boolean gdprApplies)
             {
-                setTenantConfiguration(true, url, gdprApplies);
+                log.info( "Fetch api configuration task success");
+                setTenantConfiguration(true, gdprApplies);
+                tenantConfiguration.setHost(url);
                 finishCallback.run();
             }
 
             @Override
             public void onFailure()
             {
-                setTenantConfiguration(false, null, false);
-                Log.w(TAG, "Failure onConfigurationFailure");
+                setTenantConfiguration(false, false);
+                log.error( "Fetch api configuration task failure");
                 finishCallback.run();
-                ringPublishingGDPRListener.onError(RingPublishingGDPRError.CANNOT_FETCH_TENANT_CONFIGURATION);
+                ringPublishingGDPRListener.onError(RingPublishingGDPRError.CANNOT_FETCH_TENANT_CONFIGURATION, "Fetch api configuration task failure");
             }
         });
     }
 
-    private void setTenantConfiguration(boolean success, @Nullable String url, boolean gdprApplies)
+    private void setTenantConfiguration(boolean success, boolean gdprApplies)
     {
         storage.configureGDPRApplies(gdprApplies);
         requestsState.setTenantState(success ? TenantState.SUCCESS : TenantState.FAILURE);
-        tenantConfiguration.setHost(url);
         tenantConfiguration.setGdprApplies(gdprApplies);
     }